# docs/openapi.yml
...
openapi: 3.0.3
info:
  title: Flavortown API
  version: v1
  description: You need an API key to use this! Go to your [account settings](https://flavortown.hackclub.com/kitchen?settings=1)
    to get one.
servers:
- url: https://flavortown.hackclub.com/api/v1
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Include your API key via `Authorization: Bearer YOUR_API_KEY`.
        You can find or regenerate your API key in your [account settings](https://flavortown.hackclub.com/kitchen?settings=1).'
  schemas:
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_count:
          type: integer
        next_page:
          type: integer
          nullable: true
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        repo_url:
          type: string
        demo_url:
          type: string
        readme_url:
          type: string
        ai_declaration:
          type: string
        ship_status:
          type: string
        devlog_ids:
          type: array
          items:
            type: integer
        banner_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectLink:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        link:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
    ValidationErrors:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            error: Limit cannot exceed 100
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            error: Invalid API key
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            error: Resource not found
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ValidationErrors"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
          example:
            error: rate_limited
            message: Too many requests. Please slow down.
  parameters:
    limit:
      name: limit
      in: query
      description: Number of links to return (min 1)
      required: false
      schema:
        type: integer
        minimum: 1
paths:
  "/projects":
    get:
      summary: List projects
      tags:
      - Projects
      description: 'Fetch a list of projects. Ratelimit: 5 reqs/min, 20 reqs/min if
        searching.'
      operationId: listProjects
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
        description: Page number for pagination
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
        description: Number of results per page (max 100)
      - name: query
        in: query
        required: false
        schema:
          type: string
        description: Search projects by title or description
      responses:
        '200':
          description: A paginated list of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      "$ref": "#/components/schemas/Project"
                  pagination:
                    "$ref": "#/components/schemas/Pagination"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
    post:
      summary: Create a project
      tags:
      - Projects
      description: Create a new project.
      operationId: createProject
      parameters: []
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
              - title
              - description
              properties:
                title:
                  type: string
                  description: Project title
                description:
                  type: string
                  description: Project description
                repo_url:
                  type: string
                  description: URL to the source code repository
                demo_url:
                  type: string
                  description: URL to the live demo
                readme_url:
                  type: string
                  description: URL to the README
                ai_declaration:
                  type: string
                  description: Declaration of AI tools used in this project
        required: true
  "/projects/{id}":
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    get:
      summary: Get a project
      tags:
      - Projects
      description: 'Fetch a specific project by ID. Ratelimit: 30 reqs/min.'
      operationId: getProject
      responses:
        '200':
          description: A single project
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '404':
          "$ref": "#/components/responses/NotFound"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
    patch:
      summary: Update a project
      tags:
      - Projects
      description: Update an existing project.
      operationId: updateProject
      parameters: []
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Project title
                description:
                  type: string
                  description: Project description
                repo_url:
                  type: string
                  description: URL to the source code repository
                demo_url:
                  type: string
                  description: URL to the live demo
                readme_url:
                  type: string
                  description: URL to the README
                ai_declaration:
                  type: string
                  description: Declaration of AI tools used in this project
        required: true
...
  "/users/{user_id}/projects":
    get:
      summary: List projects for a user
      tags:
      - Users
      description: Fetch a paginated list of projects belonging to a specific user.
        Use "me" as the user ID to fetch projects for the authenticated user.
      operationId: listUserProjects
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID or "me" for the authenticated user
      - name: page
        in: query
        required: false
        schema:
          type: integer
        description: Page number for pagination
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
        description: Number of results per page (max 100)
      responses:
        '200':
          description: A paginated list of the user's projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      "$ref": "#/components/schemas/Project"
                  pagination:
                    "$ref": "#/components/schemas/Pagination"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '404':
          "$ref": "#/components/responses/NotFound"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
...
  "/links/projects":
    get:
      summary: List internal project links
      tags:
      - Links
      description: 'Fetch internal project paths for all projects. Optional `limit` query param limits results.'
      operationId: listProjectLinks
      parameters:
      - "$ref": "#/components/parameters/limit"
      responses:
        '200':
          description: Project links
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_links:
                    type: array
                    items:
                      "$ref": "#/components/schemas/ProjectLink"
        '401':
          "$ref": "#/components/responses/Unauthorized"